requirements1.txt
kivy
kivymd
pyjnius
python-docx
PyPDF2
transformers
sentence-transformers